updated submit handler in addItem to prevent submit, needs review on first load
